Finished up first STL rendering
