+++ src/SUMMARY.md
@@ -21,3 +21,5 @@
     - [Enemy Racers](./swoop_enemy_racers/index.md)
     - [Engine Trails](./swoop_engine_trails/index.md)
     - [Wingtip Trails and Optimizations](./swoop_wingtip_trails/index.md)
+- [STL Viewer](./stl_viewer.md)
+    - [Load and Render STL Files](./stl_viewer_load_stl/index.md)
