+++ src/stl_viewer_load_stl/src/app.rs
@@ -2,7 +2,9 @@ use wasm_bindgen::prelude::wasm_bindgen;
 use wasm_bindgen::{JsCast, JsValue};
 use web_sys::{window, HtmlCanvasElement, KeyEvent, MouseEvent, WebGl2RenderingContext};
 
-use super::quad::Quad;
+use super::shader_stl::ShaderStl;
+use super::stl::Stl;
+use super::textures::StaticTextures;
 
 // Pull in the console.log function so we can debug things more easily
 #[wasm_bindgen]
@@ -14,7 +16,8 @@ extern "C" {
 pub struct App {
     canvas: HtmlCanvasElement,
     gl: WebGl2RenderingContext,
-    quad: Quad,
+    stl: Stl,
+    shader_stl: ShaderStl,
 }
 
 impl App {
@@ -31,15 +34,38 @@ impl App {
             panic!("No Webgl");
         }
 
-        let quad = match Quad::new(&gl) {
+        let stl = match Stl::new(&gl, include_bytes!("resources/monkey.stl")) {
             Ok(g) => g,
             Err(err) => {
-                log(&format!("Quad error {:?}", err));
-                panic!("Quad error");
+                log(&format!("Stl error {:?}", err));
+                panic!("Stl error");
             }
         };
 
-        Self { canvas, gl, quad }
+        let mut shader_stl = match ShaderStl::new(&gl) {
+            Ok(g) => g,
+            Err(err) => {
+                log(&format!("ShaderStl error {:?}", err));
+                panic!("ShaderStl error");
+            }
+        };
+
+        let textures = match StaticTextures::new(&gl) {
+            Ok(g) => g,
+            Err(err) => {
+                log(&format!("StaticTexture error {:?}", err));
+                panic!("StaticTexture error");
+            }
+        };
+
+        shader_stl.image_matcap = Some(textures.stl_matcap.clone());
+
+        Self {
+            canvas,
+            gl,
+            stl,
+            shader_stl,
+        }
     }
 
     fn check_resize(&mut self) {
@@ -55,7 +81,7 @@ impl App {
 
             self.canvas.set_width(client_width);
             self.canvas.set_height(client_height);
-            self.quad.resolution = (client_width, client_height);
+            self.shader_stl.resolution = (client_width, client_height);
 
             log(&format!("Resized to {}:{}", client_width, client_height));
         }
@@ -69,9 +95,10 @@ impl App {
 
         let now = window().unwrap().performance().unwrap().now();
         let time = (now / 1000.0) as f32;
-        self.quad.time = time;
+        self.shader_stl.time = time;
 
-        self.quad.render(&self.gl);
+        self.shader_stl.setup(&self.gl);
+        self.stl.render(&self.gl, &self.shader_stl);
     }
 
     pub fn mouse_event(&mut self, event: MouseEvent) {
