+++ src/stl_viewer_load_stl/src/resources/background.png
