+++ src/stl_viewer_load_stl/src/resources/matcap.png
