+++ src/stl_viewer_load_stl/src/textures.rs
@@ -1,0 +1,16 @@
+// Contains static textures
+
+use super::texture::{load_texture, TextureError};
+use web_sys::{WebGl2RenderingContext, WebGlTexture};
+
+pub struct StaticTextures {
+    pub stl_matcap: WebGlTexture,
+}
+
+impl StaticTextures {
+    pub fn new(gl: &WebGl2RenderingContext) -> Result<Self, TextureError> {
+        Ok(Self {
+            stl_matcap: load_texture(&gl, include_bytes!("resources/matcap.png"))?,
+        })
+    }
+}
